+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2017-09-22 - Setting to expand nav drawer menu to full page width on small screens.
 * 2017-09-22 - Check if user is logged in before showing the switched role information box.
 
 ### Release v3.2-r3